third_party/blink/web_tests/platform/generic/fast/events/touch/gesture/gesture-tap-hover-clear-expected.txt

Hover over me
Click me

Tests that hover effects from a gesture tap down can be cleared by a gesture tap or mousemove outside of the hovered element.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

See https://bugs.webkit.org/show_bug.cgi?id=103283 for details.

The Don't Click Me div should not be visible.
PASS dontClickMe.offsetTop is 0
The Don't Click Me div should be visible after a GestureShowPress on the Hover Over Me div.
PASS hoverme is hoverme
PASS dontClickMe.offsetTop is 100
The Don't Click Me div should not be visible after a GestureTap on the Click Me div.
PASS clickme is clickme
PASS dontClickMe.offsetTop is 0
The Don't Click Me div should be visible after a GestureShowPress on the Hover Over Me div.
PASS hoverme is hoverme
PASS dontClickMe.offsetTop is 100
The Don't Click Me div should not be visible after a mouse move to below the Click Me div.
PASS dontClickMe.offsetTop is 0
PASS successfullyParsed is true

TEST COMPLETE
